[fs] Add a fs segment directory and move encoding package there
