+++ index/segment/fs/encoding/checksum.go
